Update for Django 1.9
